Fix [sc-8784]: Fix sitemap connector
Fix `sync.entity` `getLastModified` function to call the method to get files from the connector only if the folders to sync is not empty. This is for preventing syncs with static folders like sitemap to upload their content several times on a same sync, and to prevent uploading the files on each synchronization.